+++ src/widgets/edit.h
@@ -220,12 +220,21 @@ typedef struct _edit_t {
    * 作为数值型编辑器时，一次增加和减少时的数值。
    */
   double step;
+  /**
+   * @property {bool_t} cancelable
+   * @annotation ["set_prop","get_prop","readable","persitent","design","scriptable"]
+   * 是否支持撤销编辑。如果为TRUE，在失去焦点之前可以撤销所有修改(恢复获得焦点之前的内容)。
+   *
+   * > * 1.一般配合keyboard的"cancel"按钮使用。
+   * > * 2.为TRUE时，如果内容有变化，会设置编辑器的状态为changed，所以此时编辑器需要支持changed状态的style。
+   */
+  bool_t cancelable;
 
   /*private*/
   uint32_t idle_id;
   uint32_t timer_id;
   text_edit_t* model;
-
+  wstr_t saved_text;
   edit_inc_value_t inc_value;
   edit_dec_value_t dec_value;
   edit_fix_value_t fix_value;
@@ -359,6 +368,17 @@ ret_t edit_set_float_limit(widget_t* widget, double min, double max, double step
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
 ret_t edit_set_readonly(widget_t* widget, bool_t readonly);
+
+/**
+ * @method edit_set_cancelable
+ * 设置编辑器是否为可撤销修改。
+ * @annotation ["scriptable"]
+ * @param {widget_t*} widget widget对象。
+ * @param {bool_t} cancelable 是否为可撤销修。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t edit_set_cancelable(widget_t* widget, bool_t cancelable);
 
 /**
  * @method edit_set_auto_fix